+++ E-Stores-UI/src/Components/Private/Seller/AddUpdateProduct.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import NavigateBtn from "../../Util/NavigateBtn";
-import { LuPlus } from "react-icons/lu";
 
 const AddUpdateProduct = () => {
   return (
