+++ E-Stores-UI/src/Components/Private/Seller/PerformanceRecord.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import DataCartBtn from "./DataCartBtn";
 import useStore from "../../Hooks/useStore";
 
 const PerformanceRecord = () => {
@@ -68,3 +67,13 @@ const PerformanceRecord = () => {
 };
 
 export default PerformanceRecord;
+
+
+export const DataCartBtn = ({name, data}) => {
+  return (
+    <div className="pt-2 flex flex-col justify-center items-center w-full hover:bg-stone-100">
+      <p>{name}</p>
+      <p className="text-2xl py-4 text-slate-500">{data}</p>
+    </div>
+  );
+};
