+++ E-Stores-UI/src/Components/Public/LoginAndRegister.jsx
@@ -3,11 +3,10 @@ import { useEffect, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { useAuth } from "../Auth/AuthProvider";
 import AxiosPrivateInstance from "../API/AxiosPrivateInstance";
-import SubmitBtn from "../Util/SubmitBtn";
 import { LuShoppingCart } from "react-icons/lu";
 import { RiShoppingBag2Line } from "react-icons/ri";
 import { HiOutlineTag } from "react-icons/hi2";
-import Input from "../Util/Input";
+import { Input, SubmitBtn } from "../Util/Forms";
 
 const Register = ({ role, isLogin }) => {
   const [email, setEmail] = useState("");
